move repository into soft HSM
