+++ components/crypto/crypto-persistence/build.gradle
@@ -28,7 +28,6 @@ dependencies {
     implementation project(":libs:crypto:crypto-core")
     implementation project(":libs:lifecycle:lifecycle")
     implementation project(':libs:virtual-node:virtual-node-info')
-    implementation project(":components:virtual-node:virtual-node-info-read-service")
 
     testImplementation "org.assertj:assertj-core:$assertjVersion"
     testImplementation "org.jetbrains.kotlin:kotlin-test:$kotlinVersion"
