+++ components/crypto/crypto-softhsm-impl/build.gradle
@@ -15,18 +15,24 @@ dependencies {
 
     implementation "com.github.ben-manes.caffeine:caffeine:$caffeineVersion"
     implementation "net.corda:corda-config-schema"
+    implementation 'net.corda:corda-db-schema'
 
     implementation project(":components:configuration:configuration-read-service")
     implementation project(":components:crypto:crypto-component-core-impl")
     implementation project(":components:crypto:crypto-hes")
     implementation project(":components:crypto:crypto-hes-core-impl")
     implementation project(":components:crypto:crypto-persistence")
+    implementation project(":components:db:db-connection-manager")
+    implementation project(":components:virtual-node:virtual-node-info-read-service")
     implementation project(":libs:cache:cache-caffeine")
     implementation project(":libs:configuration:configuration-core")
     implementation project(':libs:crypto:cipher-suite-impl')
     implementation project(':libs:crypto:crypto-core')
     implementation project(":libs:lifecycle:lifecycle")
     implementation project(':libs:utilities')
+    implementation project(':libs:virtual-node:virtual-node-info')
+
+    implementation platform("net.corda:corda-api:$cordaApiVersion")
 
     testImplementation "org.jetbrains.kotlin:kotlin-test:$kotlinVersion"
     testImplementation project(":components:crypto:crypto-component-test-utils")
