+++ components/crypto/crypto-softhsm-impl/src/main/kotlin/net/corda/crypto/softhsm/CryptoRepositoryFactory.kt
@@ -1,4 +1,4 @@
-package net.corda.crypto.persistence
+package net.corda.crypto.softhsm
 
 interface CryptoRepositoryFactory {
     /**
